refactor start forms
